mod/mpi: fix statusbar showing/hiding
The force_redraw() trick, every time a file was loaded, discarded all
messages that were shown at that time.

This behavior, for example, was the cause behind "Slide show started"
message (which is provided by the slideshow script) disappearing as soon
as the file has been loaded rather than the normal timeout.